requirements.txt
prompt_toolkit==1.0.14
whaaaaat
pillow

